Nit: Make engine symbol names more consistent
