+++ bot/http/main.ts
@@ -2,16 +2,15 @@
 // deno run -A bot/http/main.ts
 // ngrok http 8000
 
-import { postgres, serve } from "./deps.ts";
-
-import { PgStorer } from "../../storer/pg/mod.ts";
+import { serve } from "./deps.ts";
 import { mustEnv } from "../env/mod.ts";
 import { APP_PERKS } from "../env/app/mod.ts";
 import { providers } from "../env/providers/mod.ts";
 import { Registry } from "../../perks/provider/registry/mod.ts";
 import { overwrite } from "../client/client.ts";
-import { Engine } from "../../perks/engine/mod.ts";
 import { DefaultHandler } from "./handler/default/handler.ts";
+import { DefaultEngine } from "../../perks/engine/default/engine.ts";
+import { KVStorer } from "../../storer/kv/kv_storer.ts";
 
 if (import.meta.main) {
   await main();
@@ -21,17 +20,14 @@ async function main() {
   // Retrieve the required environment variables.
   const env = mustEnv();
 
-  // Create a database pool with three connections that are lazily established.
-  const pool = new postgres.Pool(env.databaseURL, 3, true);
-
   // Implement Storer class.
-  const store = new PgStorer(pool);
+  const store = new KVStorer(await Deno.openKv());
 
   // Create a Perks provider registry.
   const registry = new Registry(providers);
 
   // Create a Perks engine.
-  const engine = new Engine(store, registry);
+  const engine = new DefaultEngine(store, registry);
 
   // Overwrite the Discord Application Command.
   const { ok } = await overwrite({ ...env, app: APP_PERKS });
@@ -46,13 +42,13 @@ async function main() {
     const u = new URL(r.url);
     switch (u.pathname) {
       case env.diagnosePath: {
-        // Diagnose the database.
         const diagnosis = await store.diagnoseTables();
         return new Response(JSON.stringify(diagnosis, null, 2));
       }
 
+      // TODO: Handle env.registerPath.
+
       default: {
-        // Handle the request.
         return await handler.handle(r);
       }
     }
